Refactor: Use of extension instead of logic in IbanValidator class.
